Separate deploys and sourcemaps
